Updated components to reflect merge
