+++ public/locales/pl/mining-view.json
@@ -1,5 +1,9 @@
 {
     "floor": "Poziom",
+    "tribe": "Plemię",
+    "miners": "Górniczy",
+    "tribe-earnings": "Dochody plemienia",
+    "tribe-pool-height": "Wysokość basenu",
     "current-block-time": "Aktualny czas bloku",
     "your-reward-is": "Twoja nagroda wynosi",
     "connection-to-node-lost": "Połączenie z węzłem kopania zostało utracone. Poczekaj, aż wznowi połącznie lub uruchom go ponownie.",
