+++ src/containers/Dashboard/MiningView/MiningView.tsx
@@ -3,10 +3,12 @@ import BlockHeight from './components/BlockHeight.tsx';
 
 import Earnings from '@app/containers/Dashboard/MiningView/components/Earnings.tsx';
 import { MiningViewContainer } from './MiningView.styles.ts';
+import P2pool from '@app/containers/Dashboard/MiningView/components/P2pool.tsx';
 
 function MiningView() {
     return (
         <MiningViewContainer>
+            <P2pool />
             <Earnings />
             <BlockHeight />
             <BlockTime />
